+++ lib/read/lib/Cardano/Wallet/Read/Block.hs
@@ -6,7 +6,7 @@ The 'Block' type represents a block indexed by one of the known eras.
 It is compatible with the types from @cardano-ledger@.
 -}
 module Cardano.Wallet.Read.Block
-    ( module Cardano.Wallet.Read.Block.BHeader
+    ( module Cardano.Read.Ledger.Block.BHeader
     , module Cardano.Wallet.Read.Block.Block
     , module Cardano.Wallet.Read.Block.BlockNo
     , module Cardano.Wallet.Read.Block.HeaderHash
@@ -14,7 +14,7 @@ module Cardano.Wallet.Read.Block
     , module Cardano.Wallet.Read.Block.Txs
     ) where
 
-import Cardano.Wallet.Read.Block.BHeader
+import Cardano.Read.Ledger.Block.BHeader
     ( BHeader (..)
     , getEraBHeader
     )
