+++ lib/read/lib/Cardano/Wallet/Read/Block/BlockNo.hs
@@ -13,7 +13,7 @@ module Cardano.Wallet.Read.Block.BlockNo
 
 import Prelude
 
-import Cardano.Wallet.Read.Block.BHeader
+import Cardano.Read.Ledger.Block.BHeader
     ( BHeader (..)
     )
 import Cardano.Wallet.Read.Eras.KnownEras
