+++ lib/read/lib/Cardano/Wallet/Read/Block/SlotNo.hs
@@ -12,7 +12,7 @@ module Cardano.Wallet.Read.Block.SlotNo
 
 import Prelude
 
-import Cardano.Wallet.Read.Block.BHeader
+import Cardano.Read.Ledger.Block.BHeader
     ( BHeader (..)
     )
 import Cardano.Wallet.Read.Eras.KnownEras
